# lib/api-zod/src/generated/types/answerSaved.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerSaved {
  ok: boolean;
  /** The exact full answer read back from the database after saving. */
  persistedAnswer: string;
  /** Character length of the persisted answer. */
  savedLength: number;
  correct: boolean;
  /** Immediate numeric grade for this question (0 or 100). */
  gradePercent: number;
  explanation: string;
}